requirements.txt
aniso8601==1.2.0
appdirs==1.4.0
backports-abc==0.5
blinker==1.4
boto3==1.4.4
botocore==1.5.18
certifi==2017.1.23
cffi==1.9.1
click==6.7
cloudpickle==0.2.2
cssmin==0.2.0
cssselect==1.0.1
cymem==1.31.2
Cython==0.25.2
cytoolz==0.8.2
decorator==4.0.11
docutils==0.13.1
falcon==1.1.0
Flask==0.12
Flask-Assets==0.12
Flask-Compress==1.4.0
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-WTF==0.14.2
flexmock==0.10.2
futures==3.0.5
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.5
jmespath==0.9.1
jsmin==2.2.1
lxml==3.7.3
MarkupSafe==0.23
msgpack-python==0.4.8
murmurhash==0.26.4
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pathlib==1.0.1
pbr==1.10.0
plac==0.9.6
preshed==0.46.4
psycopg2==2.6.2
py-bcrypt==0.4
pycparser==2.17
pycurl==7.43.0
pyparsing==2.1.10
pyquery==1.2.17
python-bcrypt==0.3.2
python-dateutil==2.6.0
python-mimeparse==1.6.0
pytz==2016.10
pyzmq==16.0.2
rauth==0.7.3
redis==2.10.5
requests==2.13.0
s3transfer==0.1.10
semver==2.7.5
singledispatch==3.4.0.3
six==1.10.0
spacy==1.6.0
sputnik==0.9.3
SQLAlchemy==1.1.5
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
thinc==6.2.0
toolz==0.8.2
tornado==4.4.2
tqdm==4.11.2
ujson==1.35
uWSGI==2.0.14
webassets==0.12.1
Werkzeug==0.11.15
WTForms==2.1
zmq==0.0.0
